Nonexistant shells now panic() less.
Stop spraffing massive exceptions from both the parent and child
when the shell you're spinning up bites it by not being in the
container we're entering.

We should check what we're going to do pre-fork ideally to not
fail nastily, but we should _also_ report better than double
panic() if we do fail. This is a work in progress :)