Remove the shell loop, for different insanity.
Instead of relying on a shell at the top level of the container at all,
we instead bind dockersh into the container as /sbin/init, and
make dockersh detect if it's invoked as /sbin/init, in which case
it just logs one line to stdout then sits there till SIGQUIT or SIGINT.

This gives us a (near enough) staticly linked, predictable process
to be PID 1 and keep the container alive for the shells we later spawn
by injection into it's namespace.

This is kinda crazy, but much less gross than trying to invoke a
shell inside an unknown (and potentially hostile) container.

TODO: Add all the drop privs stuff to the /sbin/init process
too, as currently it noops as root...